+++ intelmq/bots/parsers/shadowserver/config.py
@@ -44,7 +44,7 @@ TODOs:
 import intelmq.lib.harmonization as harmonization
 
 
-def get_feed(feedname):
+def get_feed(feedname, logger):
     # TODO should this be case insensitive?
     feed_idx = {
         "Accessible-CWMP": accessible_cwmp,
@@ -53,9 +53,9 @@ def get_feed(feedname):
         "Accessible-Telnet": accessible_telnet,
         "Accessible-VNC": accessible_vnc,
         "Blacklisted-IP": blacklisted_ip,
-        "Botnet-Drone-Hadoop": botnet_drone_hadoop,
         "Compromised-Website": compromised_website,
-        "DNS-open-resolvers": dns_open_resolvers,
+        "Drone": drone,
+        "DNS-Open-Resolvers": dns_open_resolvers,
         "Microsoft-Sinkhole": microsoft_sinkhole,
         "NTP-Monitor": ntp_monitor,
         "NTP-Version": ntp_version,
@@ -68,7 +68,7 @@ def get_feed(feedname):
         "Open-MongoDB": open_mongodb,
         "Open-MSSQL": open_mssql,
         "Open-NATPMP": open_natpmp,
-        "Open-NetBIOS": open_netbios,
+        "Open-NetBIOS-Nameservice": open_netbios_nameservice,
         "Open-Netis": open_netis,
         "Open-Portmapper": open_portmapper,
         "Open-QOTD": open_qotd,
@@ -80,10 +80,22 @@ def get_feed(feedname):
         "Sandbox-URL": sandbox_url,
         "Sinkhole-HTTP-Drone": sinkhole_http_drone,
         "Spam-URL": spam_url,
-        "Ssl-Freak-Scan": ssl_freak_scan,  # Only differs in a few extra fields
-        "Ssl-Scan": ssl_scan,  # a.k.a POODLE
+        "SSL-FREAK-Vulnerable-Servers": ssl_freak_vulnerable_servers,  # Only differs in a few extra fields
+        "SSL-POODLE-Vulnerable-Servers": ssl_poodle_vulnerable_servers,
         "Vulnerable-ISAKMP": vulnerable_isakmp,
     }
+    old_feed_idx = {
+        "Botnet-Drone-Hadoop": drone,
+        "DNS-open-resolvers": dns_open_resolvers,
+        "Open-NetBIOS": open_netbios_nameservice,
+        "SSL-Freak-Scan": ssl_freak_vulnerable_servers,
+        "SSL-Scan": ssl_poodle_vulnerable_servers,
+    }
+
+    if feedname in old_feed_idx:
+        logger.warning('Deprecated feedname use. Refer to the documentation for the new name. '
+                       'Backwards compatibility will be removed in version 1.3.')
+        return old_feed_idx[feedname]
 
     return feed_idx.get(feedname)
 
@@ -211,8 +223,8 @@ open_mdns = {
         'protocol.transport': 'udp',
         'protocol.application': 'mdns',
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openmdns',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-mdns',
     }
 }
 
@@ -239,8 +251,8 @@ open_chargen = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openchargen',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-chargen',
         'protocol.application': 'chargen',
     },
 }
@@ -271,8 +283,8 @@ open_tftp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'opentftp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-tftp',
         'protocol.application': 'tftp',
     },
 }
@@ -312,8 +324,8 @@ sinkhole_http_drone = {
         # tcp.
         'protocol.transport': 'tcp',
         'classification.type': 'botnet drone',
-        'classification.taxonomy': 'Malicious Code',
-        'classification.identifier': 'botnet',
+        'classification.taxonomy': 'malicious code',
+        'classification.identifier': 'infected system',
     },
 }
 
@@ -352,8 +364,8 @@ microsoft_sinkhole = {
         'classification.type': 'botnet drone',
         'protocol.transport': 'tcp',
         'protocol.application': 'http',
-        'classification.taxonomy': 'Malicious Code',
-        'classification.identifier': 'botnet',
+        'classification.taxonomy': 'malicious code',
+        'classification.identifier': 'infected system',
     },
 }
 
@@ -392,8 +404,8 @@ open_redis = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openredis',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-redis',
         'protocol.application': 'redis',
     },
 }
@@ -423,9 +435,9 @@ open_portmapper = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openportmapper',
-        'protocol.application': 'portmapper',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-portmapper',
+        'protocol.application': 'portmap',
     },
 }
 
@@ -467,8 +479,8 @@ open_ipmi = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openipmi',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-ipmi',
         'protocol.application': 'ipmi',
         'protocol.transport': 'udp',
     },
@@ -498,8 +510,8 @@ open_qotd = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openqotd',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-qotd',
         'protocol.application': 'qotd',
     },
 }
@@ -537,8 +549,8 @@ open_ssdp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openssdp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-ssdp',
         'protocol.application': 'ssdp',
     },
 }
@@ -567,9 +579,9 @@ open_snmp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
+        'classification.taxonomy': 'vulnerable',
         'protocol.application': 'snmp',
-        'classification.identifier': 'opensnmp',
+        'classification.identifier': 'open-snmp',
     },
 }
 
@@ -601,8 +613,8 @@ open_mssql = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openmssql',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-mssql',
         'protocol.application': 'mssql',
     },
 }
@@ -639,14 +651,14 @@ open_mongodb = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openmongodb',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-mongodb',
         'protocol.application': 'mongodb',
     },
 }
 
 # https://www.shadowserver.org/wiki/pmwiki.php/Services/Open-NetBIOS
-open_netbios = {
+open_netbios_nameservice = {
     'required_fields': [
         ('time.source', 'timestamp', add_UTC_to_timestamp),
         ('source.ip', 'ip'),
@@ -668,8 +680,8 @@ open_netbios = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'opennetbios',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-netbios',
         'protocol.application': 'netbios',
     },
 }
@@ -706,8 +718,8 @@ open_elasticsearch = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openelasticsearch',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-elasticsearch',
         'protocol.application': 'elasticsearch',
     },
 }
@@ -734,8 +746,8 @@ dns_open_resolvers = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'opendns',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'dns-open-resolver',
         'protocol.application': 'dns',
     },
 }
@@ -757,14 +769,14 @@ ntp_monitor = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openntp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'ntp-monitor',
         'protocol.application': 'ntp',
     },
 }
 
 # https://www.shadowserver.org/wiki/pmwiki.php/Services/Ssl-Freak-Scan
-ssl_freak_scan = {
+ssl_freak_vulnerable_servers = {
     'required_fields': [
         ('time.source', 'timestamp', add_UTC_to_timestamp),
         ('source.ip', 'ip'),
@@ -779,7 +791,7 @@ ssl_freak_scan = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
+        'classification.taxonomy': 'vulnerable',
         'classification.identifier': 'SSL-FREAK',
         'protocol.application': 'https',
     },
@@ -787,7 +799,7 @@ ssl_freak_scan = {
 
 
 # https://www.shadowserver.org/wiki/pmwiki.php/Services/Ssl-Scan
-ssl_scan = {
+ssl_poodle_vulnerable_servers = {
     'required_fields': [
         ('time.source', 'timestamp', add_UTC_to_timestamp),
         ('source.ip', 'ip'),
@@ -802,8 +814,8 @@ ssl_scan = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'SSL-Poodle',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'SSL-POODLE',
         'protocol.application': 'https',
     },
 }
@@ -825,14 +837,14 @@ open_memcached = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
-        'classification.identifier': 'openmemcached',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-memcached',
         'protocol.application': 'memcached',
     },
 }
 
 # https://www.shadowserver.org/wiki/pmwiki.php/Services/Botnet-Drone-Hadoop
-botnet_drone_hadoop = {
+drone = {
     'required_fields': [
         ('time.source', 'timestamp', add_UTC_to_timestamp),
         ('source.ip', 'ip'),
@@ -863,8 +875,8 @@ botnet_drone_hadoop = {
     ],
     'constant_fields': {
         'classification.type': 'botnet drone',
-        'classification.taxonomy': 'Malicious Code',
-        'classification.identifier': 'botnet',
+        'classification.taxonomy': 'malicious code',
+        'classification.identifier': 'infected system',
     },
 }
 
@@ -891,10 +903,10 @@ open_xdmcp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.taxonomy': 'Vulnerable',
+        'classification.taxonomy': 'vulnerable',
         'protocol.application': 'xdmcp',
         # 'feed.url': 'https://www.shadowserver.org/wiki/pmwiki.php/Services/Open-XDMCP',
-        'classification.identifier': 'openxdmcp',
+        'classification.identifier': 'open-xdmcp',
     },
 }
 
@@ -924,6 +936,7 @@ compromised_website = {
     ],
     'constant_fields': {
         'classification.type': 'compromised',
+        'classification.taxonomy': 'intrusions',
         'classification.identifier': 'compromised-website',
     },
 }
@@ -951,8 +964,9 @@ open_natpmp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'opennatpmp',
-        'protocol.application': 'nat-pmp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-natpmp',
+        'protocol.application': 'natpmp',
     },
 }
 
@@ -975,7 +989,8 @@ open_netis = {
     'constant_fields': {
         'protocol.transport': 'udp',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'opennetis',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-netis',
     },
 }
 
@@ -1024,7 +1039,8 @@ ntp_version = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'openntpversion',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'ntp-version',
         'protocol.application': 'ntp',
     },
 }
@@ -1046,7 +1062,8 @@ sandbox_url = {
     ],
     'constant_fields': {
         'classification.type': 'malware',
-        'classification.identifier': 'sandboxurl',
+        'classification.taxonomy': 'malicious code',
+        'classification.identifier': 'sandbox-url',
     },
 }
 
@@ -1073,7 +1090,8 @@ spam_url = {
     ],
     'constant_fields': {
         'classification.type': 'spam',
-        'classification.identifier': 'spamurl',
+        'classification.taxonomy': 'abusive content',
+        'classification.identifier': 'spam-url',
     },
 }
 
@@ -1108,7 +1126,8 @@ vulnerable_isakmp = {
     ],
     'constant_fields': {
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'openike',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-ike',
     }
 }
 
@@ -1149,7 +1168,8 @@ accessible_rdp = {
         'protocol.transport': 'tcp',
         'protocol.application': 'rdp',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'openrdp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-rdp',
     },
 }
 
@@ -1226,7 +1246,8 @@ open_ldap = {
     'constant_fields': {
         'protocol.application': 'ldap',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'openldap',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-ldap',
     }
 }
 
@@ -1250,7 +1271,8 @@ blacklisted_ip = {
     ],
     'constant_fields': {
         'classification.type': 'blacklist',
-        'classification.identifier': 'blacklisted',
+        'classification.taxonomy': 'other',
+        'classification.identifier': 'blacklisted-ip',
     }
 }
 
@@ -1276,7 +1298,8 @@ accessible_telnet = {
         'protocol.transport': 'tcp',
         'protocol.application': 'telnet',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'opentelnet',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-telnet',
     }
 }
 
@@ -1312,7 +1335,8 @@ accessible_cwmp = {
     'constant_fields': {
         'protocol.application': 'cwmp',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'opencwmp',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'open-cwmp',
     }
 }
 
@@ -1337,6 +1361,7 @@ accessible_vnc = {
         'protocol.transport': 'tcp',
         'protocol.application': 'vnc',
         'classification.type': 'vulnerable service',
-        'classification.identifier': 'accessiblevnc',
+        'classification.taxonomy': 'vulnerable',
+        'classification.identifier': 'accessible-vnc',
     }
 }