+++ NEWS.md
@@ -9,6 +9,9 @@ See the changelog for a full list of changes.
 - `intelmqctl start` prints bot's error messages in stderr if it failed to start.
 - `intelmqctl check` checks if all keys in the packaged defaults.conf are present in the current configuration.
 
+### Configuration
+A new harmonization type `JSONDict` has been added specifically for the `extra` field. It is highly recommended to change the type of this field.
+
 1.0.0 Stable release
 --------------------
 ### Configuration